backend/requirements.txt
# Flask Backend Requirements
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# ML & Computer Vision
# Use tensorflow-cpu for free-tier deployments (no GPU needed)
tensorflow-cpu==2.15.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.24.3
scikit-learn==1.4.0
scikit-image>=0.21.0
h5py>=3.10.0
mediapipe>=0.10.9

# Utilities
python-dotenv==1.0.0
werkzeug==3.0.1
gunicorn==21.2.0
eventlet==0.36.1
